modularize, refactor, commented and cleaned up
