+++ app/helpers/input.js
@@ -1,0 +1,5 @@
+const userInput = process.argv;
+const command = process.argv[2];
+
+module.exports.userInput = userInput;
+module.exports.command = command;
