+++ log.txt
@@ -31,3 +31,31 @@ movie-this => {"Title":"My Neighbor Totoro","Year":"1988","Rating":"G","RottenTo
 movie-this => {"Title":"My Neighbor Totoro","Year":"1988","Rating":"G","RottenTomatoes":"94%","Country":"Japan","Language":"Japanese","Plot":"When two girls move to the country to be near their ailing mother, they have adventures with the wondrous forest spirits who live nearby.","Actors":"Noriko Hidaka, Chika Sakamoto, Shigesato Itoi, Sumi Shimamoto"},
 movie-this => {"Title":"Bokassa 1er, empereur de Françafrique","Year":"2011","Rating":"N/A","RottenTomatoes":"N/A","Country":"France","Language":"N/A","Plot":"N/A","Actors":"Henri Maïdou, Ange-Félix Patassé, Francis-Albert Ouakanga, Jean-Paul Benoît"},
 concert-this => {"Band":"Bokassa","Venue":"Valdebebas - IFEMA","City":"Madrid","Country":"Spain","Date":"03/05/2019"},
+movie-this => {"Title":"Casa Blanca","Year":"2015","Rating":"N/A","RottenTomatoes":"N/A","Country":"Poland, Mexico, Cuba","Language":"Spanish","Plot":"Casa Blanca is a small fishing village on the Gulf of Havana. Nelsa (76) and her son Vladimir (37), who has Down syndrome, share a tiny room in an overcrowded multi-family building. ...","Actors":"N/A"},
+movie-this => {"Title":"The Cat Returns","Year":"2002","Rating":"G","RottenTomatoes":"90%","Country":"Japan","Language":"Japanese, English","Plot":"After helping a cat, a 17-year-old girl finds herself involuntarily engaged to a cat prince in a magical world where her only hope of freedom lies with a dapper cat statuette come to life.","Actors":"Chizuru Ikewaki, Yoshihiko Hakamada, Aki Maeda, Takayuki Yamada"},
+movie-this => {"Title":"Alien","Year":"1979","Rating":"R","RottenTomatoes":"97%","Country":"UK, USA","Language":"English","Plot":"After a space merchant vessel perceives an unknown transmission as a distress call, its landing on the source moon finds one of the crew attacked by a mysterious lifeform, and they soon realize that its life cycle has merely begun.","Actors":"Tom Skerritt, Sigourney Weaver, Veronica Cartwright, Harry Dean Stanton"},
+movie-this => {"Title":"Alien","Year":"1979","Rating":"R","RottenTomatoes":"97%","Country":"UK, USA","Language":"English","Plot":"After a space merchant vessel perceives an unknown transmission as a distress call, its landing on the source moon finds one of the crew attacked by a mysterious lifeform, and they soon realize that its life cycle has merely begun.","Actors":"Tom Skerritt, Sigourney Weaver, Veronica Cartwright, Harry Dean Stanton"},
+movie-this => {"Title":"Alien","Year":"1979","Rating":"R","RottenTomatoes":"97%","Country":"UK, USA","Language":"English","Plot":"After a space merchant vessel perceives an unknown transmission as a distress call, its landing on the source moon finds one of the crew attacked by a mysterious lifeform, and they soon realize that its life cycle has merely begun.","Actors":"Tom Skerritt, Sigourney Weaver, Veronica Cartwright, Harry Dean Stanton"},
+concert-this => {"Band":"Slayer","Venue":"Brisbane Riverstage ","City":"Brisbane","Country":"Australia","Date":"07/03/2019"},
+concert-this => {"Band":"Slayer","Venue":"Brisbane Riverstage ","City":"Brisbane","Country":"Australia","Date":"07/03/2019"},
+spotify-this-song => {"Artist":"IZ*ONE","Song":"La Vie en Rose","OnSpotify":"https://open.spotify.com/track/1fRaznZsAB2i63TtmYv2Hc","Album":"COLOR*IZ","Year":"29/10/2018"},
+spotify-this-song => {"Artist":"Adele","Song":"Hello","OnSpotify":"https://open.spotify.com/track/4sPmO7WMQUAf45kwMOtONw","Album":"25","Year":"24/06/2016"},
+movie-this => {"Title":"Jaw","Year":"2005","Rating":"N/A","RottenTomatoes":"N/A","Country":"USA","Language":"English","Plot":"A jawless, strip club bouncer (Jeff Maschi) fights freak show bouts for low level celebs while recalling his glory days as a legitimate boxer with a mean right hook and a chin like a concrete block.","Actors":"Frank Coleman, Jeff Maschi"},
+movie-this => {"Title":"Jaws","Year":"1975","Rating":"PG","RottenTomatoes":"97%","Country":"USA","Language":"English","Plot":"When a killer shark unleashes chaos on a beach resort, it's up to a local sheriff, a marine biologist, and an old seafarer to hunt the beast down.","Actors":"Roy Scheider, Robert Shaw, Richard Dreyfuss, Lorraine Gary"},
+movie-this => {"Title":"Jaws","Year":"1975","Rating":"PG","RottenTomatoes":"97%","Country":"USA","Language":"English","Plot":"When a killer shark unleashes chaos on a beach resort, it's up to a local sheriff, a marine biologist, and an old seafarer to hunt the beast down.","Actors":"Roy Scheider, Robert Shaw, Richard Dreyfuss, Lorraine Gary"},
+spotify-this-song => {"Artist":"Audioslave","Song":"I Am the Highway","OnSpotify":"https://open.spotify.com/track/3bwAS8cN7Xtzvv5UurJOOX","Album":"Audioslave","Year":"17/11/2002"},
+spotify-this-song => {"Artist":"Audioslave","Song":"I Am the Highway","OnSpotify":"https://open.spotify.com/track/3bwAS8cN7Xtzvv5UurJOOX","Album":"Audioslave","Year":"17/11/2002"},
+concert-this => {"Band":"Bokassa","Venue":"Valdebebas - IFEMA","City":"Madrid","Country":"Spain","Date":"03/05/2019"},
+concert-this => {"Band":"Kiss","Venue":"End of the Road World Tour - Vancouver, Canada","City":"Vancouver","Country":"Canada","Date":"31/01/2019"},
+spotify-this-song => {"Artist":"Adele","Song":"Hello","OnSpotify":"https://open.spotify.com/track/4sPmO7WMQUAf45kwMOtONw","Album":"25","Year":"24/06/2016"},
+movie-this => {"Title":"TRON","Year":"1982","Rating":"PG","RottenTomatoes":"71%","Country":"USA","Language":"English","Plot":"A computer hacker is abducted into the digital world and forced to participate in gladiatorial games where his only chance of escape is with the help of a heroic security program.","Actors":"Jeff Bridges, Bruce Boxleitner, David Warner, Cindy Morgan"},
+movie-this => {"Title":"Jaws","Year":"1975","Rating":"PG","RottenTomatoes":"97%","Country":"USA","Language":"English","Plot":"When a killer shark unleashes chaos on a beach resort, it's up to a local sheriff, a marine biologist, and an old seafarer to hunt the beast down.","Actors":"Roy Scheider, Robert Shaw, Richard Dreyfuss, Lorraine Gary"},
+spotify-this-song => {"Artist":"Audioslave","Song":"I Am the Highway","OnSpotify":"https://open.spotify.com/track/3bwAS8cN7Xtzvv5UurJOOX","Album":"Audioslave","Year":"17/11/2002"},
+concert-this => {"Band":"Bokassa","Venue":"Valdebebas - IFEMA","City":"Madrid","Country":"Spain","Date":"03/05/2019"},
+spotify-this-song => {"Artist":"Adele","Song":"Hello","OnSpotify":"https://open.spotify.com/track/4sPmO7WMQUAf45kwMOtONw","Album":"25","Year":"24/06/2016"},
+spotify-this-song => {"Artist":"Various Artists","Song":"ASDF - Original Mix","OnSpotify":"https://open.spotify.com/track/1GkhxYsJvsrvE4dNkzvIC6","Album":"New Blood Of Bass, Vol. 4","Year":"09/09/2013"},
+concert-this => {"Band":"Bokassa","Venue":"Valdebebas - IFEMA","City":"Madrid","Country":"Spain","Date":"03/05/2019"},
+concert-this => {"Band":"Bokassa","Venue":"Valdebebas - IFEMA","City":"Madrid","Country":"Spain","Date":"03/05/2019"},
+spotify-this-song => {"Artist":"IZ*ONE","Song":"La Vie en Rose","OnSpotify":"https://open.spotify.com/track/1fRaznZsAB2i63TtmYv2Hc","Album":"COLOR*IZ","Year":"29/10/2018"},
+movie-this => {"Title":"TRON","Year":"1982","Rating":"PG","RottenTomatoes":"71%","Country":"USA","Language":"English","Plot":"A computer hacker is abducted into the digital world and forced to participate in gladiatorial games where his only chance of escape is with the help of a heroic security program.","Actors":"Jeff Bridges, Bruce Boxleitner, David Warner, Cindy Morgan"},
+movie-this => {"Title":"Jaws","Year":"1975","Rating":"PG","RottenTomatoes":"97%","Country":"USA","Language":"English","Plot":"When a killer shark unleashes chaos on a beach resort, it's up to a local sheriff, a marine biologist, and an old seafarer to hunt the beast down.","Actors":"Roy Scheider, Robert Shaw, Richard Dreyfuss, Lorraine Gary"},
